poem_verse_app/scripts/requirements.txt
# Python dependencies for utility scripts
# Install with: pip install -r requirements.txt

matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0
